+++ pkg/cep_ent/schema/invoke_model_order.go
@@ -18,10 +18,10 @@ func (InvokeModelOrder) Fields() []ent.Field {
 		field.Int64("api_token_id").StructTag(`json:"api_token"`).Comment("API Token ID"),
 		field.Enum("invoke_type").StructTag(`json:"invoke_type"`).Default(string(enums.UnKnownInvokeType)).GoType(enums.WebInvokeType).Comment("调用类型"),
 		field.Int("invoke_times").StructTag(`json:"invoke_times"`).Default(0).Comment("调用次数"),
-		field.Int("input_token_cost").StructTag(`json:"input_token_cost"`).Default(0).Comment("输入token消耗"),
-		field.Int("output_token_cost").StructTag(`json:"output_token_cost"`).Default(0).Comment("输出token消耗"),
-		field.Int("input_cep_cost").StructTag(`json:"input_cep_cost"`).Default(0).Comment("输入cep消耗"),
-		field.Int("output_cep_cost").StructTag(`json:"output_cep_cost"`).Default(0).Comment("输出cep消耗"),
+		field.Int64("input_token_cost").StructTag(`json:"input_token_cost"`).Default(0).Comment("输入token消耗"),
+		field.Int64("output_token_cost").StructTag(`json:"output_token_cost"`).Default(0).Comment("输出token消耗"),
+		field.Int64("input_cep_cost").StructTag(`json:"input_cep_cost"`).Default(0).Comment("输入cep消耗"),
+		field.Int64("output_cep_cost").StructTag(`json:"output_cep_cost"`).Default(0).Comment("输出cep消耗"),
 	}
 }
